New11

+++ src/Prime.cpp
@@ -4,7 +4,7 @@ void Prime::setValue(int val) {
 	this->number = val;
 }
 bool Prime::isValid() {
-	for (size_t i = 2; i < this->number; ++i){
+	for (int i = 2; i < this->number; ++i){
 		if (this->number % i == 0) {
 			return false;
 		}
@@ -12,7 +12,7 @@ bool Prime::isValid() {
 	return true;
 }
 bool Prime::testPrime(int val) {
-	for (size_t i = 2; i < val; ++i) {
+	for (int i = 2; i < val; ++i) {
 		if (val % i == 0) {
 			return false;
 		}
@@ -22,7 +22,7 @@ bool Prime::testPrime(int val) {
 int Prime::countBetween(Prime& simple) {
 	int count{ 0 };
 
-	for (size_t i = (simple.number > this->number) ? this->number : simple.number + 1, max = (simple.number > this->number) ? simple.number : this->number; i < max; ++i){
+	for (int i = ((simple.number > this->number) ? this->number : simple.number + 1), max = ((simple.number > this->number) ? simple.number : this->number); i < max; ++i){
 		if (testPrime(i)) {
 			count++;
 		}
